move escapeLikeString function to database package
