+++ app/api/currentuser/search_for_available_groups.go
@@ -5,10 +5,10 @@ import (
 	"net/http"
 	"strings"
 	"unicode/utf8"
-	"unsafe"
 
 	"github.com/go-chi/render"
 
+	"github.com/France-ioi/AlgoreaBackend/app/database"
 	"github.com/France-ioi/AlgoreaBackend/app/service"
 )
 
@@ -106,7 +106,7 @@ func (srv *Service) searchForAvailableGroups(w http.ResponseWriter, r *http.Requ
 		Where("group_pending_requests.type IN ('join_request', 'invitation')").
 		SubQuery()
 
-	escapedSearchString := escapeLikeString(searchString, '|')
+	escapedSearchString := database.EscapeLikeString(searchString, '|')
 	query := srv.Store.Groups().
 		Select(`
 			groups.id,
@@ -134,36 +134,4 @@ func (srv *Service) searchForAvailableGroups(w http.ResponseWriter, r *http.Requ
 
 	render.Respond(w, r, convertedResult)
 	return service.NoError
-}
-
-// escapeLikeStringBackslash escapes string with backslashes the given escape character.
-// This escapes the contents of a string (provided as string)
-// by adding the escape character before percent signs (%), and underscore signs (_).
-func escapeLikeString(v string, escapeCharacter byte) string {
-	pos := 0
-	buf := make([]byte, len(v)*3)
-
-	for i := 0; i < len(v); i++ {
-		c := v[i]
-		switch c {
-		case escapeCharacter:
-			buf[pos] = escapeCharacter
-			buf[pos+1] = escapeCharacter
-			pos += 2
-		case '%':
-			buf[pos] = escapeCharacter
-			buf[pos+1] = '%'
-			pos += 2
-		case '_':
-			buf[pos] = escapeCharacter
-			buf[pos+1] = '_'
-			pos += 2
-		default:
-			buf[pos] = c
-			pos++
-		}
-	}
-
-	result := buf[:pos]
-	return *(*string)(unsafe.Pointer(&result)) // nolint:gosec
 }
